+++ prowler/providers/aws/services/mq/mq_broker_active_deployment_mode/mq_broker_active_deployment_mode.metadata.json
@@ -1,7 +1,7 @@
 {
   "Provider": "aws",
-  "CheckID": "mq_activemq_broker_active_standby_mode",
-  "CheckTitle": "ActiveMQ brokers should be configured in active/standby mode.",
+  "CheckID": "mq_broker_active_deployment_mode",
+  "CheckTitle": "Apache ActiveMQ brokers should be configured in active/standby mode.",
   "CheckType": [
     "Software and Configuration Checks/Industry and Regulatory Standards/NIST 800-53 Controls"
   ],
@@ -10,8 +10,8 @@
   "ResourceIdTemplate": "arn:aws:mq:region:account-id:broker:broker-id",
   "Severity": "low",
   "ResourceType": "AwsAmazonMQBroker",
-  "Description": "Ensure Amazon MQ ActiveMQ brokers are configured in active/standby mode for high availability and fault tolerance.",
-  "Risk": "ActiveMQ brokers not configured in active/standby mode lack high availability, increasing the risk of downtime and data loss during failures.",
+  "Description": "Ensure Amazon MQ Apache ActiveMQ brokers are configured in active/standby mode for high availability and fault tolerance.",
+  "Risk": "Apache ActiveMQ brokers not configured in active/standby mode lack high availability, increasing the risk of downtime and data loss during failures.",
   "RelatedUrl": "https://docs.aws.amazon.com/amazon-mq/latest/developer-guide/welcome.html",
   "Remediation": {
     "Code": {
@@ -21,7 +21,7 @@
       "Terraform": "https://www.trendmicro.com/cloudoneconformity/knowledge-base/aws/MQ/deployment-mode.html"
     },
     "Recommendation": {
-      "Text": "Ensure Amazon MQ ActiveMQ brokers use active/standby deployment mode for high availability and fault tolerance.",
+      "Text": "Ensure Amazon MQ Apache ActiveMQ brokers use active/standby deployment mode for high availability and fault tolerance.",
       "Url": "https://docs.aws.amazon.com/amazon-mq/latest/developer-guide/amazon-mq-broker-architecture.html"
     }
   },
